refactor: group common code inside `org.eclipse.kura.nm.common` package
