+++ kura/org.eclipse.kura.nm/src/main/java/org/eclipse/kura/nm/common/NMDeviceEnable.java
@@ -10,7 +10,7 @@
  * Contributors:
  *  Eurotech
  *******************************************************************************/
-package org.eclipse.kura.nm.configuration;
+package org.eclipse.kura.nm.common;
 
 public enum NMDeviceEnable {
