+++ kura/org.eclipse.kura.nm/src/main/java/org/eclipse/kura/nm/common/NetworkProperties.java
@@ -10,7 +10,7 @@
  * Contributors:
  *  Eurotech
  *******************************************************************************/
-package org.eclipse.kura.nm.configuration;
+package org.eclipse.kura.nm.common;
 
 import java.util.ArrayList;
 import java.util.List;
