+++ kura/test/org.eclipse.kura.nm.test/src/test/java/org/eclipse/kura/nm/configuration/writer/DhcpServerConfigWriterTest.java
@@ -27,7 +27,7 @@ import org.apache.commons.io.FileUtils;
 import org.eclipse.kura.KuraException;
 import org.eclipse.kura.linux.net.dhcp.DhcpServerManager;
 import org.eclipse.kura.linux.net.dhcp.DhcpServerTool;
-import org.eclipse.kura.nm.configuration.NetworkProperties;
+import org.eclipse.kura.nm.common.NetworkProperties;
 import org.junit.After;
 import org.junit.Test;
 import org.mockito.MockedStatic;
